+++ src/global.d.ts
@@ -1,6 +1,3 @@
 declare module 'execa'
 declare module 'is-ci'
-declare module 'pkg-dir'
 declare module 'slash'
-declare module 'tempy'
-declare module 'read-pkg'