+++ src/installer/is.ts
@@ -1,7 +1,7 @@
 import { huskyIdentifier } from './getScript'
 
 export function isHusky(data: string): boolean {
-  // v0.14 and prior used #husky as an identifier.
+  // Husky v0.14 and prior used #husky as an identifier.
   // Just in case some previous hooks weren't correctly uninstalled,
   // and for a better transition this will allow v0.15+ to uninstall them as well.
   const previousHuskyIdentifier = '#husky'