+++ src/upgrader/index.ts
@@ -1,12 +1,12 @@
-import * as fs from 'fs'
-import * as path from 'path'
-import * as readPkg from 'read-pkg'
+import fs from 'fs'
+import path from 'path'
+import readPkg from 'read-pkg'
 
-interface IHookMap {
+interface HookMap {
   [key: string]: string
 }
 
-const hookList: IHookMap = {
+const hookList: HookMap = {
   applypatchmsg: 'applypatch-msg',
   commitmsg: 'commit-msg',
   postapplypatch: 'post-applypatch',
@@ -28,33 +28,41 @@ const hookList: IHookMap = {
   update: 'update'
 }
 
-export default function upgrade(cwd: string) {
+export default function upgrade(cwd: string): void {
   const pkgFile = path.join(cwd, 'package.json')
   if (fs.existsSync(pkgFile)) {
     const pkg = readPkg.sync({ cwd, normalize: false })
 
     console.log(`husky > upgrading ${pkgFile}`)
 
-    // Don't overwrite pkg.husky if it exists
+    // Don't overwrite 'husky' field if it exists
     if (pkg.husky) {
       return console.log(
         `husky field in package.json isn't empty, skipping automatic upgrade`
       )
     }
 
-    const hooks: IHookMap = {}
+    const hooks: HookMap = {}
 
-    // Loop trhough hooks and move them to husky.hooks
-    Object.keys(hookList).forEach(name => {
-      const script = pkg.scripts[name]
-      if (script) {
-        delete pkg.scripts[name]
-        const newName = hookList[name]
-        hooks[newName] = script.replace(/\bGIT_PARAMS\b/g, 'HUSKY_GIT_PARAMS')
-        console.log(`moved scripts.${name} to husky.hooks.${newName}`)
+    // Find hooks in package.json 'scripts' field
+    Object.keys(hookList).forEach(
+      (name: string): void => {
+        if (pkg.scripts) {
+          const script = pkg.scripts[name]
+          if (script) {
+            delete pkg.scripts[name]
+            const newName = hookList[name]
+            hooks[newName] = script.replace(
+              /\bGIT_PARAMS\b/g,
+              'HUSKY_GIT_PARAMS'
+            )
+            console.log(`moved scripts.${name} to husky.hooks.${newName}`)
+          }
+        }
       }
-    })
+    )
 
+    // Move found hooks to 'husky.hooks' field
     if (Object.keys(hooks).length) {
       pkg.husky = { hooks }
     } else {